requirements.txt
tweepy==4.14.0
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
twilio==8.13.0
openai>=1.0.0
